requirements.txt
Flask==2.3.2
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
pillow==10.0.0
psycopg2-binary
Flask-RESTful
Flask-RESTPlus
opencv-python-headless==4.8.0.74
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.26.4
python-dotenv==0.21.0
gunicorn==20.1.0
thefuzz==0.20.0
